Fixed some problems:  - Fixed FC when ImageCache runs extremely long  - Moved some error messages to resources  - Translated some strings  - Integrated new Support-Libraray v7 21.1.1 which caused problems with the theme  - Pinned the library versions of the support libraries

+++ ttrssreader/src/main/java/org/ttrssreader/net/JavaJSONConnector.java
@@ -65,9 +65,8 @@ public class JavaJSONConnector extends JSONConnector {
 			con.setRequestProperty("Content-Length", Integer.toString(outputBytes.length));
 
 			// Timeouts
-			int timeoutSocket = (int) ((Controller.getInstance().lazyServer()) ? 15 * Utils.MINUTE : 10 * Utils
-					.SECOND);
-			con.setReadTimeout(timeoutSocket);
+			long timeoutSocket = (Controller.getInstance().lazyServer()) ? 15 * Utils.MINUTE : 10 * Utils.SECOND;
+			con.setReadTimeout((int) timeoutSocket);
 			con.setConnectTimeout((int) (8 * Utils.SECOND));
 
 			// HTTP-Basic Authentication
